terminada la parte de admin -funcionando-

+++ ecommerce-web/src/main/java/web/ecommerce/tpfinal/ecommerce_web/clasesDominio/Producto.java
@@ -11,7 +11,7 @@ import javax.persistence.OneToOne;
 public class Producto {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 	
 	private String nombre;
 	
@@ -24,7 +24,7 @@ public class Producto {
 	}
 
 
-	public Producto(String nombre, Fabricante fabricante, float precio, int idProducto) {
+	public Producto(String nombre, Fabricante fabricante, float precio, Long idProducto) {
 		this.nombre = nombre;
 		this.fabricante = fabricante;
 		this.precio = precio;
@@ -54,10 +54,10 @@ public class Producto {
 	public void setPrecio(float precio) {
 		this.precio = precio;
 	}
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(int idProducto) {
+	public void setId(Long idProducto) {
 		this.id = idProducto;
 	}
